refactor(app_view): split open_app_window into window.rs
Lifts the window-construction entry point out of mod.rs into
`app_view::window`, re-exported as `app_view::open_app_window`
so main.rs's `use app_view::{open_app_window, AppView,
WindowInit};` continues to resolve without change.

Moves: `pub fn open_app_window` (~135 lines) + the
`TRAFFIC_LIGHT_POSITION_PX` const that backs its
`traffic_light_position` titlebar option.

The const stays shared with `AppView::open_settings_window` in
mod.rs (the Settings window uses the same traffic-light
position so the two windows feel uniform). Exposed via
`pub(super)` and re-imported into mod.rs by name.

`open_app_window` calls the still-private `AppView::connect_to`
and reads the still-private `AppView::profile_store` field from
its sibling module — that works without visibility widening
because Rust's privacy rule lets descendant modules see the
parent's private items (window.rs is a descendant of the
app_view module that defines AppView).

Verified: `cargo clippy --all-targets -- -D warnings` and
`cargo test` both pass (63 tests, unchanged).
